Bug Fix
Noticed that the delta altitude wasn't changing.  I think that's because the initial value it was reading was NAN which was causing it to be stuck at zero.  I added a one-time extra read prior to collecting altitude data to bypass the problem of reading a funky value the first time.  May be a better way by adding delay or something, but I think it is always the first meas that comes back NAN (noticed it in the terminal once).  Seems to work now.